Dbg board: subsystem organization
Did relative placement for all subsystems (processor, regulator, usb- hub, etc.)

It is unlikely that all components on this board can fit into a small footprint due to the size and number of components. Some further squeezing and optimization can be done, but part of the board will need to be larger to expose vertically accessed ports (debug).