DBG-placement on board
It's pretty large; about 3.5x2.5. Lots of ports and test points that need to be accessible.  Test points should probably be moved around so the text is more readable.

Everything important to touch is top-side, and pushed to the exterior of the board.

The boundaries of the board push out to the side such that the physical platform should be pretty stable (in terms of balance w/o a bunch of cables coming off of it.